Move notebooks to one directory
