+++ .gitignore
@@ -2,3 +2,4 @@ venv/
 __pycache__/
 .ipynb_checkpoints/
 checkpoint.pth
+.vscode
